Our job: fill in the 5 methods from UserInterface. But check this out - I'm going rogue - I'm only going to implement 2 of them: getUsername() and getRoles().
getUsername()

First getUsername()... is super-unimportant. Just return any unique user string you want - a username, an email, a uuid, a funny, but unique joke - whatever. This is only used to show you who is logged in when you're debugging.

In our app, our users won't have a username - but they will have an email. Add a private {dollar_sign}email property then just use that in getUsername(): return {dollar_sign}this->email.

And add a setter for that method.

We'll eventually need that to create users.