getCredentials(): Build the Form
If the user is trying to login, our new task is to fetch the username & password and return them.

Since we built a form, let's let the form do the work for us.

Normally in a controller, we call {dollar_sign}this->createForm() to build the form:

398 lines  vendor/symfony/symfony/src/Symfony/Bundle/FrameworkBundle/Controller/Controller.php
... lines 1 - 38
abstract class Controller implements ContainerAwareInterface
{
... lines 41 - 274
  /**
  * Creates and returns a Form instance from the type of the form.
  *
  * @param string|FormTypeInterface     The built type of the form
  * @param mixed                        The initial data for the form
  * @param array                     Options for the form
  *
  * @return Form
  */
  protected function createForm({dollar_sign}type, {dollar_sign}data = null, array {dollar_sign}options = array())
  {
    return {dollar_sign}this->container->get('form.factory')->create({dollar_sign}type, {dollar_sign}data, {dollar_sign}options);
  }
... lines 288 - 396

In reality, this grabs the form.factory service and calls create() on it.

Dependency Inject form.factory (FormFactory)

So how can we create a form in the authenticator? Use dependency injection to inject the form.factory service.

Add a __construct() method with a formFactory argument.

Now, I like to type-hint my arguments, so let's just guess at the service's class name and see if there's one called FormFactory. Yep, there's even a FormFactoryInterface!

That's probably what we want. I'll press Option+Enter and select 'Initialize Fields' to set that property for me.
If you're still getting used to dependency injection and that all happened too fast, don't worry. We know we want to inject the form.factory service, so I guessed its class for the type-hint, which is optional. You can always find your terminal and run:

bin/console debug:container form.factory
to find out the exact class to use for the type-hint. We will also register this as a service in services.yml in a minute.