When Authentication Fails? getLoginUrl()
That's it! Authenticators are always these three methods.

But, what happens if authentication fails? Where should we send the user? And what about when the login is successful?

When authentication fails, we need to redirect the user back to the login form. That will happen automatically - we just need to fill in getLoginUrl() so the system knows where that is.

But to do that, we'll need the router service. Once again, go back to the top and add another constructor argument for the router. To be super cool, you can type-hint with the RouterInterface.
Use the Option+Enter shortcut again to set up that property.
Down in getLoginUrl(), return {dollar_sign}this->router->generate('security_login')