In security.yml, you already have a providers section - as in 'user providers'. Delete the in_memory stuff and replace it with our_users: that's a totally meaningless machine name - it could be anything. But below that, say entity and set it to { class: AppBundle\Entity\User, property: email }.
The property part is not something we care about right now, but we will use and talk about it later.

But yea, that's it! Go back to /login. Right now, I am not logged in. But try logging in again.

It's alive!!! We can finally surf around the site and stay logged in. Cool.

Custom User Provider

In your app, if you're not loading users from the database, then you'll need to create a custom user provider class that implements UserProviderInterface. Check out the official docs(http://symfony.com/doc/current/cookbook/security/custom_provider.html) in this case.