Check this out: let's fail authentication with a bad password.
Ok: I noticed two things. First, we have an error:

Invalid credentials.
Great! But second, the form is not pre-filled with the email address I just used. Hmm.

Behind the scenes, the authenticator communicates to your SecurityController by storing things in the session. That's what the security.authentication_utils helps us with:

 36 lines  src/AppBundle/Controller/SecurityController.php
... lines 1 - 8
class SecurityController extends Controller
{
... lines 11 - 13
  public function loginAction()
  {
    {dollar_sign}authenticationUtils = {dollar_sign}this->get('security.authentication_utils');
    // get the login error if there is one
    {dollar_sign}error = {dollar_sign}authenticationUtils->getLastAuthenticationError();
... lines 20 - 34
  }
}
Hold command and open getLastAuthenticationError(). Ultimately, this reads a Security::AUTHENTICATION_ERROR string key from the session.

And the same is true for fetching the last username, or email in our case: it reads a key from the session.

Here's the deal: the login form is automatically setting the authentication error to the session for us. But, it is not setting the last username on the session... because it doesn't really know where to look for it.