I just found out that giving everyone the same password - iliketurtles - is apparently not a great security system. Let's give each user their own password. Again, in your security setup, you might not be responsible for storing and checking passwords. Skip this if it doesn't apply to you.
In User, add a private  that will eventually store the encoded password. Give it the @ORM\Column annotation.

Now, remember the three methods from UserInterface that we left blank?

It's finally their time to shine. In getPassword(), return .

But keep getSalt() blank: we're going to use the bcrypt algorithm, which has a built-in mechanism to salt passwords.

Use the 'Code'->'Generate' menu to generate the setter for password