In AppBundle, create a new directory called Doctrine and a new class called HashPasswordListener.
If this is your first Doctrine listener, welcome! They're pretty friendly. Here's the idea: we'll create a function that Doctrine will call whenever any entity is inserted or updated. That'll let us to do some work before that happens.

Implement an EventSubscriber interface and then use Command+N or the 'Code'->'Generate' menu, select 'Implement Methods' and choose the one method: getSubscribedEvents().
In here, return an array with prePersist and preUpdate.

These are two event names that Doctrine makes available. prePersist is called right before an entity is originally inserted. preUpdate is called right before an entity is updated.