Now, on to encoding that password.
Symfony comes with a built-in service that's really good at encoding passwords. It's called security.password_encoder and if you looked it up on debug:container, its class is UserPasswordEncoder. We'll need that, so add a __construct() function and type-hint a single argument with UserPasswordEncoder {dollar_sign}passwordEncoder. I'll hit Option+Enter and select 'Initialize Fields' to save me some time.

In a minute, we'll register this as a service.

Down below, add {dollar_sign}encoded = {dollar_sign}this->passwordEncoder->encodePassword() and pass it the User - which is {dollar_sign}entity - and the plain-text password: {dollar_sign}entity->getPlainPassword(). Finish it with {dollar_sign}entity->setPassword({dollar_sign}encoded).

That's it: we are encoded!