Denying access is great... but we still have a User class that gives every user the same, hardcoded role: ROLE_USER
et's say that in our system, we're going to give different users different roles. How do we do that? Simple! Just create a private  property that's an array. Give it the @ORM\Column annotation and set its type to json_array.
Tip
json_array type is deprecated since Doctrine 2.6, you should use json instead.

This is really cool because the  property will hold an array of roles, but when we save, Doctrine will automatically json_encode that array and store it in a single field. When we query, it'll json_decode that back to the array. What this means is that we can store an array inside a single column, without ever worrying about the JSON encode stuff.