There's really only 2 things you can do with security:
1.Deny access
2.Find out who is logged in
To show that off, find newAction(). Let's update the flash message to include the email address of the current user.

Surround the string with sprintf and add a %s placeholder right in the middle. How can you find out who's logged in? Yep, it's {dollar_sign}this->getUser(). And that returns - wait for it - our User object. Which allows us to use any methods on it, like getUsername(). That method rerturns our username's email if you remember.

Nice! Now go and create a sea monster. Fill out all the fields... and... eureka!

The Secret Behind getUser()

But you know I hate secrets: I want to know what that getUser() method really does. So hold Command and click to see that method.

The important piece is that the user comes from a service called security.token_storage:

 398 lines  vendor/symfony/symfony/src/Symfony/Bundle/FrameworkBundle/Controller/Controller.php
... lines 1 - 38
abstract class Controller implements ContainerAwareInterface
{
... lines 41 - 317
  /**
  * Get a user from the Security Token Storage.
  *
  * @return mixed
  *
  * @throws \LogicException If SecurityBundle is not available
  *
  * @see TokenInterface::getUser()
  */
  protected function getUser()
  {
      throw new \LogicException('The SecurityBundle is not registered in your application.');
    }
    if (null === {dollar_sign}token = {dollar_sign}this->container->get('security.token_storage')->getToken()) {
      return;
    }
      // e.g. anonymous authentication
      return;
    }
    return ;
  }
... lines 344 - 396
}
So if you ever need the User object in a service, this is how to get it.

But, it takes a couple of steps: getToken() gives you a pretty unimportant object, except for the fact that you can call getUser() on it.

The Creepy anon.

But, there's a pitfall: if you are anonymous - so, not logged in - getUser() does not return null, as you might expect: it returns a string: anon.. I know - it's super weird. So, if you ever do fetch the user object directly via this service, check to make sure getUser() returns an object. If it doesn't, the user isn't logged in.