The one other place where you'll need to fetch the User is inside Twig. In fact, let's talk about security in general in Twig. Open up base.html.twig.
Earlier, we already showed how to check for a role in Twig: it's via the is_granted() function:

53 lines  app/Resources/views/base.html.twig
<html>
... lines 3 - 13
  <body>
... lines 15 - 19
    <header class='header'>
... lines 21 - 22
      <ul class='navi'>
... line 24
        {% if is_granted('ROLE_USER') %}
          <li><a href='{{ path('security_logout') }}'>Logout</a></li>
        {% else %}
          <li><a href='{{ path('security_login') }}'>Login</a></li>
        {% endif %}
      </ul>
    </header>
... lines 32 - 50
  </body>
</html>

It's easy: it works exactly the same as in the controller.

So, how do we get the user object? To find out - open up the homepage template. If the User is logged in, let's welcome them by email.

Open the print tag and say app.user ? app.user.email : 'Aquanauts'

That app variable is the one global variable you get in Symfony. And one of its super-powers is to give you the current User object if there is one, or null if the user isn't logged in.