Next, this will be a nice, normal form situation. So click the Form directory, open the new menu, and create a new Symfony Form. Call it UserRegistrationForm. Brilliant! Delete the extra getName() method that's super not needed in Symfony 3. Now, bind the form to User, with ->setDefaults() and a data_class option to set User::class.
Then, we do need a password field, but think about it: the property we want to set on User is not actually the password property. We need to set plainPassword. Add this. It'll be a password type. But, if you want the user to type the password twice, use a RepeatedType. Then, in the third argument, pass the real type with type set to PasswordType::class.

That'll render two password boxes. And if the values don't match, validation will automatically fail.