Registration is working, but it's missing validation.
Since the form is bound to the User class, that is where our annotation rules should live. First, you need the use statement for the annotations. We added validation earlier in Genus. So, you can either copy this use statement, grab it from the documentation, or do what I do: cheat by saying use, auto-completing an annotation I know exists - like NotBlank, deleting the last part, and adding the normal as Assert alias.

We obviously want email to be NotBlank. We also want email to be a valid email address. For plainPassword, that should also not be blank.

Pretty simple.

Ok, go back, keep the form blank, and submit. Nice validation errors.